+++ activerecord/test/models/author.rb
@@ -19,6 +19,7 @@ class Author < ActiveRecord::Base
   has_many :comments_containing_the_letter_e, :through => :posts, :source => :comments
   has_many :comments_with_order_and_conditions, -> { order('comments.body').where("comments.body like 'Thank%'") }, :through => :posts, :source => :comments
   has_many :comments_with_include, -> { includes(:post) }, :through => :posts, :source => :comments
+  has_many :post_comments, :through => :post, :source => :comments
 
   has_many :first_posts
   has_many :comments_on_first_posts, -> { order('posts.id desc, comments.id asc') }, :through => :first_posts, :source => :comments